Partie_1/requirements.txt
six==1.16.0
git+https://github.com/dpkp/kafka-python.git
pyspark==3.5.1
redis==5.0.1
pymongo==4.6.1
pytz==2024.1
neo4j==5.14.0
psycopg2-binary